% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_poesia.R
\docType{data}
\name{poesia_df}
\alias{poesia_df}
\title{Base de datos de poesía}
\format{
Un data frame con X filas y Y columnas:
\describe{
\item{autor}{Nombre del autor del poema}
\item{titulo}{Título del poema}
\item{texto}{Contenido del poema}
\item{año}{Año de publicación}
}
}
\source{
Datos recolectados de fuentes literarias públicas.
}
\usage{
poesia_df
}
\description{
Este dataset contiene información sobre varios poetas y sus poemas.
}
\examples{
data(poesia_df)
head(poesia_df)
}
\keyword{datasets}
